Move more code out of SDK, HOB and table functions
